+++ client/index.scss
@@ -26,109 +26,130 @@ body {
 
 #homepage {
     display: flex;
+    flex-direction: column;
 
-    #sidebar {
-        width: 150px;
-        margin-top: 3vh;
-        padding-left: 20px;
-
-        button {
-            margin-bottom: 20px;
-            border: none;
-            background-color: white;
-            color: black;
-            padding: 0px;
-            font-size: 13px;
-            outline: none;
-            cursor: pointer;
-        }
-    }
-
-    #box-container {
+    #random-question-container{
+        height: 300px;
         display: flex;
-        flex-wrap: wrap;
-        width: calc(100vw - 150px);
-
-        .box {
-            margin-left: 1.5vw;
-            margin-right: 1.5vw;
-            margin-top: 2vh;
-            margin-bottom: 2vh;
-            width: 25vw;
-            height: 200px;
-            padding: 15px;
-
-            .title {
-                margin-bottom: 30px;
-                font-size: 18px;
-                text-align: center;
-                border: 1px solid black;
-                margin-left: 4vw;
-                margin-right: 4vw;
-            }
-
-            .questions {
-                height: 150px;
-                font-size: 13px;
-                line-height: 17px;
-                margin-left: 5vw;
-
-                .question-link {
-                    text-decoration: underline;
-                    color: black;
-                }
-            }
-
-            .description {
-                height: 150px;
-                font-size: 13px;
-                line-height: 18px;
+        justify-content: center;
+        #question-column-container{
+            display: flex;
+            flex-direction: column;
+            justify-content: center;
+            font-size: 20px;
+            line-height: 40px;
+            #randomQuestionLink {
+                color: black;
+                text-decoration: underline;
             }
         }
     }
 
-    #difficulty-container {
+
+    #homepage-bottom-container {
         display: flex;
-        flex-wrap: wrap;
-        width: calc(100vw - 150px);
-        .difficulty-box {
-            margin-left: 1.5vw;
-            margin-right: 1.5vw;
-            margin-top: 2vh;
-            margin-bottom: 2vh;
-            width: 25vw;
-            height: 200px;
-            padding: 15px;
-
-            .title {
-                margin-bottom: 30px;
-                font-size: 18px;
-                text-align: center;
-                border: 1px solid black;
-                margin-left: 4vw;
-                margin-right: 4vw;
-            }
-
-            .questions {
-                height: 150px;
+        #sidebar {
+            width: 150px;
+            margin-top: 3vh;
+            padding-left: 20px;
+    
+            button {
+                margin-bottom: 20px;
+                border: none;
+                background-color: white;
+                color: black;
+                padding: 0px;
                 font-size: 13px;
-                line-height: 17px;
-                margin-left: 5vw;
+                outline: none;
+                cursor: pointer;
+            }
+        }
+        #box-container {
+            display: flex;
+            flex-wrap: wrap;
+            width: calc(100vw - 150px);
 
-                .question-link {
-                    text-decoration: underline;
-                    color: black;
+            .box {
+                margin-left: 1.5vw;
+                margin-right: 1.5vw;
+                margin-top: 2vh;
+                margin-bottom: 2vh;
+                width: 25vw;
+                height: 200px;
+                padding: 15px;
+
+                .title {
+                    margin-bottom: 30px;
+                    font-size: 18px;
+                    text-align: center;
+                    border: 1px solid black;
+                    margin-left: 4vw;
+                    margin-right: 4vw;
+                }
+
+                .questions {
+                    height: 150px;
+                    font-size: 13px;
+                    line-height: 17px;
+                    margin-left: 5vw;
+
+                    .question-link {
+                        text-decoration: underline;
+                        color: black;
+                    }
+                }
+
+                .description {
+                    height: 150px;
+                    font-size: 13px;
+                    line-height: 18px;
                 }
             }
-
         }
 
+        #difficulty-container {
+            display: flex;
+            flex-wrap: wrap;
+            width: calc(100vw - 150px);
+
+            .difficulty-box {
+                margin-left: 1.5vw;
+                margin-right: 1.5vw;
+                margin-top: 2vh;
+                margin-bottom: 2vh;
+                width: 25vw;
+                height: 200px;
+                padding: 15px;
+
+                .title {
+                    margin-bottom: 30px;
+                    font-size: 18px;
+                    text-align: center;
+                    border: 1px solid black;
+                    margin-left: 4vw;
+                    margin-right: 4vw;
+                }
+
+                .questions {
+                    height: 150px;
+                    font-size: 13px;
+                    line-height: 17px;
+                    margin-left: 5vw;
+
+                    .question-link {
+                        text-decoration: underline;
+                        color: black;
+                    }
+                }
+            }
+        }
     }
 }
 
 #app {
     height: 100vh;
     width: 100vw;
+
     .navbar {
         width: 100vw;
         height: 30px;
